permissions

+++ src/features/ticket/actions/delete-ticket.ts
@@ -11,19 +11,25 @@ import { getAuthOrRedirect } from "@/features/auth/queries/get-auth-or-redirect"
 import { isOwner } from "@/features/auth/utils/is-owner";
 import { prisma } from "@/lib/prisma";
 import { ticketsPath } from "@/paths";
+import { getTicketPermissions } from "../permissions/get-ticket-permissions";
 
 export const deleteTicket = async (id: string) => {
   const { user } = await getAuthOrRedirect();
   try {
-    if (id) {
-      const ticket = await prisma.ticket.findUnique({
-        where: {
-          id,
-        },
-      });
-      if (!ticket || !isOwner(user, ticket)) {
-        return toActionState("ERROR", "Not authorized");
-      }
+    const ticket = await prisma.ticket.findUnique({
+      where: {
+        id,
+      },
+    });
+    if (!ticket || !isOwner(user, ticket)) {
+      return toActionState("ERROR", "Not authorized");
+    }
+    const permissions = await getTicketPermissions({
+      organizationId: ticket.organizationId,
+      userId: user?.id,
+    });
+    if (!permissions.canDeleteTicket) {
+      return toActionState("ERROR", "Not authorized");
     }
     await prisma.ticket.delete({ where: { id } });
   } catch (error) {
